Fix BytesToReadable, use uint64_t
Resolves 32-bit builds limited to 4GB accurate file size reporting

Removes the warning within ImGuiFileDialog

+++ src/core/util/string/string.cc
@@ -69,7 +69,7 @@ BuildPath(
 
 std::string
 BytesToReadable(
-	size_t bytes,
+	uint64_t bytes,
 	ByteConversionFlags flags
 )
 {
